+++ lib/res/res.dart
@@ -1,0 +1,3 @@
+export './event_bus.dart';
+export './app_event.dart';
+export './subscription.dart';